refactor(core): remove outbox_models shim
