--- apps/backend/app/core/outbox_models.py
@@ -1,8 +1,0 @@
-from __future__ import annotations
-
-import app.models.outbox as _outbox  # legacy source
-
-OutboxEvent = _outbox.OutboxEvent
-OutboxStatus = _outbox.OutboxStatus
-
-__all__ = ["OutboxEvent", "OutboxStatus"]
